+++ src/grants.rs
@@ -7,17 +7,18 @@
 //! incoming messages to determine whether they have sufficient privileges to
 //! undertake the message's action(s).
 
-mod grant;
+mod verify;
 
 use base64ct::{Base64UrlUnpadded, Encoding};
+use chrono::{DateTime, Utc};
 use serde::{Deserialize, Serialize};
 
-pub use self::grant::{Conditions, Grant, GrantData, Publication, RequestData, RevocationData};
 use crate::protocols::{GRANT_PATH, PROTOCOL_URI, REQUEST_PATH, REVOCATION_PATH};
 use crate::provider::MessageStore;
-use crate::records::{RecordsFilter, Write};
+use crate::records::{DelegatedGrant, RecordsFilter, Write};
+use crate::serde::rfc3339_micros;
 use crate::store::RecordsQueryBuilder;
-use crate::{Interface, Method, Result, forbidden};
+use crate::{Interface, Method, Result, forbidden, unexpected};
 
 /// Fetches the grant specified by `grant_id`.
 pub async fn fetch_grant(owner: &str, grant_id: &str, store: &impl MessageStore) -> Result<Grant> {
@@ -39,7 +40,7 @@ pub async fn fetch_grant(owner: &str, grant_id: &str, store: &impl MessageStore)
         return Err(forbidden!("missing grant data"));
     };
     let grant_bytes = Base64UrlUnpadded::decode_vec(grant_enc)?;
-    let grant: grant::GrantData = serde_json::from_slice(&grant_bytes)?;
+    let grant: GrantData = serde_json::from_slice(&grant_bytes)?;
 
     Ok(Grant {
         id: write.record_id.clone(),
@@ -83,6 +84,157 @@ pub async fn fetch_scope(owner: &str, write: &Write, store: &impl MessageStore) 
 
         _ => Err(forbidden!("invalid `protocol_path`")),
     }
+}
+
+/// [`Grant`] holds permission grant information during the process of
+/// verifying an incoming message's authorization.
+#[derive(Clone, Debug, Default, Deserialize, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct Grant {
+    /// The ID of the permission grant — the record ID message.
+    pub id: String,
+
+    /// The grantor of the permission.
+    pub grantor: String,
+
+    /// The grantee of the permission.
+    pub grantee: String,
+
+    /// The date at which the grant was given.
+    #[serde(serialize_with = "rfc3339_micros")]
+    pub date_granted: DateTime<Utc>,
+
+    /// The grant's descriptor.
+    #[serde(flatten)]
+    pub data: GrantData,
+}
+
+impl TryFrom<&Write> for Grant {
+    type Error = crate::Error;
+
+    fn try_from(write: &Write) -> Result<Self> {
+        let permission_grant = write.encoded_data.clone().unwrap_or_default();
+        let grant_data = serde_json::from_str(&permission_grant)
+            .map_err(|e| unexpected!("issue deserializing grant: {e}"))?;
+
+        Ok(Self {
+            id: write.record_id.clone(),
+            grantor: write.authorization.signer().unwrap_or_default(),
+            grantee: write.descriptor.recipient.clone().unwrap_or_default(),
+            date_granted: write.descriptor.date_created,
+            data: grant_data,
+        })
+    }
+}
+
+impl TryFrom<&DelegatedGrant> for Grant {
+    type Error = crate::Error;
+
+    fn try_from(delegated: &DelegatedGrant) -> Result<Self> {
+        let bytes = Base64UrlUnpadded::decode_vec(&delegated.encoded_data)?;
+        let grant_data = serde_json::from_slice(&bytes)
+            .map_err(|e| unexpected!("issue deserializing grant: {e}"))?;
+
+        Ok(Self {
+            id: delegated.record_id.clone(),
+            grantor: delegated.authorization.signer()?,
+            grantee: delegated.descriptor.recipient.clone().unwrap_or_default(),
+            date_granted: delegated.descriptor.date_created,
+            data: grant_data,
+        })
+    }
+}
+
+/// The [`GrantData`] data structure holds information related to a permission
+/// grant.
+///
+/// The grant is issued as a [`Write`] message with the base64 URL-encoded
+/// [`GrantData`] structure in the `encoded_data` field.
+#[derive(Clone, Debug, Default, Deserialize, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct GrantData {
+    /// Describes the intended grant use.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub description: Option<String>,
+
+    /// The CID of permission request. This is optional as grants may be issued
+    /// without being requested.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub request_id: Option<String>,
+
+    /// The date-time when grant expires.
+    #[serde(serialize_with = "rfc3339_micros")]
+    pub date_expires: DateTime<Utc>,
+
+    /// Specifies whether grant is delegated or not. When set to `true`, the
+    /// `granted_to` property acts as the `granted_to` within the scope of the
+    /// grant.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub delegated: Option<bool>,
+
+    /// The scope of access permitted by the grant.
+    pub scope: Scope,
+
+    /// Specifies an conditions that must be met when the grant is used.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub conditions: Option<Conditions>,
+}
+
+/// [`Conditions`] contains conditions set on the parent `Grant`.
+#[derive(Clone, Debug, Default, Deserialize, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct Conditions {
+    /// Indicates whether a message written with the invocation of a permission
+    /// must, may, or must not be marked as public. If unset, it is optional to
+    /// make the message public.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub publication: Option<Publication>,
+}
+
+/// Condition for publication of a message.
+#[derive(Clone, Debug, Default, Deserialize, Serialize, PartialEq, Eq)]
+pub enum Publication {
+    /// The message must be marked as public.
+    #[default]
+    Required,
+
+    /// The message may be marked as public.
+    Prohibited,
+}
+
+/// The [`RequestData`] data structure holds information related to a permission
+/// grant request.
+///
+/// The request is made using a [`Write`] message with the base64 URL-encoded
+/// [`RequestData`] structure in the `encoded_data` field.
+#[derive(Clone, Debug, Default, Deserialize, Serialize)]
+pub struct RequestData {
+    /// If the grant is a delegated grant or not. If `true`, `granted_to` will
+    /// be able to act as the `granted_by` within the scope of this grant.
+    pub delegated: bool,
+
+    /// Optional string that communicates what the grant would be used for.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub description: Option<String>,
+
+    /// The scope of the allowed access.
+    pub scope: Scope,
+
+    /// Optional conditions that must be met when the grant is used.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub conditions: Option<Conditions>,
+}
+
+/// The [`RevocationData`] data structure holds information related to a permission
+/// grant revocation.
+///
+/// The revocation is saved as a [`Write`] message with the base64 URL-encoded
+/// [`RevocationData`] structure in the `encoded_data` field.
+#[derive(Clone, Debug, Deserialize, Serialize)]
+pub struct RevocationData {
+    /// Optional string that communicates the details of the revocation.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub description: Option<String>,
 }
 
 /// The `Scope` enum specifies the interface-specific scope of a permission
